[Fix] Update tsconfig [Change] Move tests files location.
