+++ tsconfig.build.json
@@ -1,5 +1,7 @@
 
 {
   "extends": "./tsconfig",
-  "exclude": ["example"]
+  "exclude": [
+    "example"
+  ]
 }
